Update Folder Structure + Code + Readme
