+++ lib/features/auth/sign_in/blocs/sign_in_state.dart
@@ -1,9 +1,11 @@
-part of 'auth_sign_in_bloc.dart';
+part of 'sign_in_bloc.dart';
 
 @immutable
 abstract class AuthSignInState {}
 
 final class AuthSignInInitial extends AuthSignInState {}
+
+final class AuthSignInLoading extends AuthSignInState {}
 
 final class AuthSignInSucces extends AuthSignInState {}
