+++ lib/features/auth/sign_in/repository/sign_in_repo.dart
@@ -4,6 +4,6 @@ class SignInRepo {
   final FirebaseAuth auth = FirebaseAuth.instance;
 
   Future<void> signIn({required String email, required String password}) async {
-    await auth.createUserWithEmailAndPassword(email: email, password: password);
+    await auth.signInWithEmailAndPassword(email: email, password: password);
   }
 }
