+++ lib/features/auth/sign_up/blocs/sign_up_event.dart
@@ -1,4 +1,4 @@
-part of 'auth_sign_up_bloc.dart';
+part of 'sign_up_bloc.dart';
 
 @immutable
 abstract class AuthSignUpEvent {}
