+++ lib/features/next_match/pages/next_match_details.dart
