+++ lib/main_development.dart
@@ -4,10 +4,10 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:soccer_club_app/auth_sign_in_repo/bloc/auth_sign_in_bloc.dart';
-import 'package:soccer_club_app/auth_sign_in_repo/sign_in_repo.dart';
-import 'package:soccer_club_app/auth_sign_up_repo/bloc/auth_sign_up_bloc.dart';
-import 'package:soccer_club_app/auth_sign_up_repo/sign_up_repo.dart';
+import 'package:soccer_club_app/features/auth/sign_in/blocs/sign_in_bloc.dart';
+import 'package:soccer_club_app/features/auth/sign_up/blocs/sign_up_bloc.dart';
+import 'package:soccer_club_app/features/auth/sign_in/repository/sign_in_repo.dart';
+import 'package:soccer_club_app/features/auth/sign_up/repository/sign_up_repo.dart';
 import 'package:soccer_club_app/core/theme/app_theme.dart';
 import 'package:soccer_club_app/l10n/l10n.dart';
 import 'package:soccer_club_app/routes/routes.dart';
