+++ lib/models/avatar_model.dart
