+++ lib/pages/favourite/player.dart
@@ -5,7 +5,7 @@ import 'package:soccer_club_app/core/color/app_color.dart';
 import 'package:soccer_club_app/core/constant/icons.dart';
 import 'package:soccer_club_app/core/extention/builder_context_extension.dart';
 import 'package:soccer_club_app/core/utils/size_utils.dart';
-import 'package:soccer_club_app/data/models/avatar_model.dart';
+import 'package:soccer_club_app/models/avatar_model.dart';
 import 'package:soccer_club_app/l10n/l10n.dart';
 import 'package:soccer_club_app/routes/routes.dart';
 import 'package:soccer_club_app/widgets/button.dart';
